[update]コードを機能ごとにモジュール分割
- 質問追加機能を `add_question_handler.js` に分割
- テキストフォーム処理を `text_form_handler.js` に分割
- 質問フォーム生成処理を `question_form.js` に分割
- 質問一括送信処理を `submit_handler.js` に分割
- `image_to_text_and_generate_questions.js` を他のモジュールを利用する形に修正
